pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "dash>=2.9.3",
    "dash-bootstrap-components>=1.7.1",
    "email-validator>=2.2.0",
    "faker>=37.3.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "plotly>=6.1.2",
    "psycopg2-binary>=2.9.10",
    "schedule>=1.2.2",
    "sendgrid>=6.12.3",
    "werkzeug>=3.1.3",
    "xlsxwriter>=3.2.3",
]
